backend/requirements.txt
flask==3.0.0
openai>=1.12.0
python-dotenv==1.0.0
werkzeug==3.0.1
flask-cors==4.0.0
pydub==0.25.1
httpx>=0.25.0